+++ wayland-scanner/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Use wrapper type implementing `Sync` instead of `static mut`s.
+- Add headerless xml file parsing possibility for `parse` function.
 
 ## 0.31.1 -- 2024-01-29